LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssPositionValue-invalid-expected.txt


FAIL Constructing a CSSPositionValue with a keyword throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a keyword throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a keyword throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a double throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a double throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a double throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a unitless zero throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a unitless zero throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a unitless zero throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a string length throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a string length throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a string length throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a number CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a number CSSUnitValue throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a number CSSUnitValue throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a time CSSUnitValue throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a time CSSUnitValue throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a time CSSUnitValue throws a TypeError Can't find variable: CSSPositionValue
FAIL Constructing a CSSPositionValue with a CSSMathValue of angle type throws a TypeError assert_throws_js: function "() => new CSSPositionValue(value, CSS.px(0))" threw object "ReferenceError: Can't find variable: CSSPositionValue" ("ReferenceError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL Updating CSSPositionValue.x with a CSSMathValue of angle type throws a TypeError Can't find variable: CSSPositionValue
FAIL Updating CSSPositionValue.y with a CSSMathValue of angle type throws a TypeError Can't find variable: CSSPositionValue
